requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.6.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.9
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==2.4.4
idna==2.10
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
JPype1==1.0.2
keyext @ file:///home/user1/server/keyword-extraction
konlpy==0.5.2
krwordrank==1.0.3
lxml==4.5.2
MarkupSafe==1.1.1
numpy==1.19.2
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
PySocks==1.7.1
regex==2020.10.15
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.3.19
textrank @ git+https://github.com/lovit/textrank/@fb06f6d451b3da2238e97869305ac31fc16a76ae
threadpoolctl==2.1.0
tqdm==4.50.2
tweepy==3.9.0
urllib3==1.25.10
Werkzeug==1.0.1
